refactor: classify in more detail
