+++ app/src/main/AndroidManifest.xml
@@ -15,10 +15,10 @@
         android:theme="@style/Theme.Map"
         tools:targetApi="31">
         <activity
-            android:name=".view.KakaoMapActivity"
+            android:name=".view.kakaomap.KakaoMapActivity"
             android:exported="false" />
         <activity
-            android:name=".view.SearchWindowActivity"
+            android:name=".view.search.SearchWindowActivity"
             android:exported="false" />
         <activity
             android:name=".view.MainActivity"
