+++ app/src/main/java/campus/tech/kakao/map/repository/search/SearchRepository.kt
@@ -1,8 +1,8 @@
-package campus.tech.kakao.map.repository
+package campus.tech.kakao.map.repository.search
 
 import campus.tech.kakao.map.BuildConfig
-import campus.tech.kakao.map.model.Place
-import campus.tech.kakao.map.model.SearchKeyword
+import campus.tech.kakao.map.model.search.Place
+import campus.tech.kakao.map.model.search.SearchKeyword
 import kotlinx.coroutines.coroutineScope
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
