+++ app/src/main/java/campus/tech/kakao/map/repository/search/StoreInfoDBHelper.kt
@@ -1,10 +1,10 @@
-package campus.tech.kakao.map.repository
+package campus.tech.kakao.map.repository.search
 
 import android.content.Context
 import android.database.sqlite.SQLiteDatabase
 import android.database.sqlite.SQLiteOpenHelper
 import android.provider.BaseColumns
-import campus.tech.kakao.map.model.SearchKeywordEntry
+import campus.tech.kakao.map.model.search.SearchKeywordEntry
 
 private const val SQL_SEARCH_KEYWORD_CREATE_ENTRIES =
     "CREATE TABLE ${SearchKeywordEntry.TABLE_NAME} (" +
