+++ app/src/main/java/campus/tech/kakao/map/view/MainActivity.kt
@@ -2,9 +2,9 @@ package campus.tech.kakao.map.view
 
 import android.content.Intent
 import android.os.Bundle
-import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
 import campus.tech.kakao.map.R
+import campus.tech.kakao.map.view.kakaomap.KakaoMapActivity
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
