+++ app/src/main/java/campus/tech/kakao/map/view/search/SearchResultsAdapter.kt
@@ -1,10 +1,10 @@
-package campus.tech.kakao.map.view
+package campus.tech.kakao.map.view.search
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import campus.tech.kakao.map.databinding.SearchResultItemBinding
-import campus.tech.kakao.map.model.Place
+import campus.tech.kakao.map.model.search.Place
 
 class SearchResultsAdapter(
     private val searchResults: List<Place>,
