+++ app/src/main/java/campus/tech/kakao/map/viewmodel/kakaomap/KakaoMapViewModelFactory.kt
@@ -1,8 +1,8 @@
-package campus.tech.kakao.map.viewmodel
+package campus.tech.kakao.map.viewmodel.kakaomap
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import campus.tech.kakao.map.repository.LastPositionRepository
+import campus.tech.kakao.map.repository.kakaomap.LastPositionRepository
 
 class KakaoMapViewModelFactory(
     private val lastPositionRepository: LastPositionRepository
