+++ app/src/main/res/layout/activity_main.xml
@@ -4,6 +4,6 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.KakaoMapActivity">
+    tools:context=".view.kakaomap.KakaoMapActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
